Refactor payloads to permit adding owners to the blob meta structure without pushing it from the client
