+++ bin/amphora/src/amphora/server/handlers/delete.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use apikit::{auth::UserIdentity, reject::InternalServerError};
 
-use interface::{message as msg, StorageNode};
+use interface::StorageNode;
 
 use warp::reply;
 
@@ -15,7 +15,5 @@ pub async fn delete<N: StorageNode>(
         .await
         .map_err(InternalServerError::from)?;
 
-    Ok(apikit::reply::json(&msg::MessageResponse {
-        message: String::from("OK"),
-    }))
+    Ok(apikit::reply::message("OK"))
 }
