+++ bin/amphora/src/amphora/server/handlers/fsync.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use apikit::{auth::UserIdentity, reject::InternalServerError};
 
-use interface::{message::MessageResponse, StorageNode};
+use interface::StorageNode;
 
 use warp::reply;
 
@@ -15,7 +15,5 @@ pub async fn fsync<N: StorageNode>(
         .await
         .map_err(InternalServerError::from)?;
 
-    Ok(apikit::reply::json(&MessageResponse {
-        message: String::from("OK"),
-    }))
+    Ok(apikit::reply::message("OK"))
 }
