+++ bin/amphora/src/amphora/server/handlers/health.rs
@@ -1,11 +1,7 @@
 use std::sync::Arc;
 
-use interface::{message::storage_node as msg, StorageNode};
-
-use warp::reply;
+use interface::StorageNode;
 
 pub async fn health<N: StorageNode>(_node: Arc<N>) -> Result<impl warp::Reply, warp::Rejection> {
-    Ok(reply::json(&msg::MessageResponse {
-        message: String::from("healthy"),
-    }))
+    Ok(apikit::reply::message("healthy"))
 }
