+++ bin/amphora/src/amphora/server/handlers/version.rs
@@ -1,11 +1,9 @@
-use interface::message::VersionResponse;
+use protocol::VersionResponse;
 
 use warp::reply;
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 
 pub async fn version() -> Result<reply::Response, warp::Rejection> {
-    Ok(apikit::reply::json(&VersionResponse {
-        version: VERSION.to_string(),
-    }))
+    Ok(apikit::reply::json(&VersionResponse::new(VERSION)))
 }
