+++ bin/menmosd/Cargo.toml
@@ -48,6 +48,7 @@ ipnetwork = "0.17"
 log = "0.4"
 log4rs = "1"
 mime = "0.3"
+protocol = {path = "../../crates/protocol"}
 rapidquery = {path = "../../crates/rapidquery", features=["bitvec_span"]}
 serde = {version = "1", features = ["derive"]}
 serde_json = "1"
