+++ bin/menmosd/src/menmosd/server/handlers/admin/health.rs
@@ -1,9 +1,3 @@
-use interface::message as msg;
-
-use warp::reply;
-
 pub async fn health() -> Result<impl warp::Reply, warp::Rejection> {
-    Ok(reply::json(&msg::MessageResponse {
-        message: String::from("healthy"),
-    }))
+    Ok(apikit::reply::message("healthy"))
 }
