+++ bin/menmosd/src/menmosd/server/handlers/admin/version.rs
@@ -1,12 +1,11 @@
 use apikit::auth::UserIdentity;
-use interface::message::VersionResponse;
+
+use protocol::VersionResponse;
 
 use warp::reply;
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 
 pub async fn version(_user: UserIdentity) -> Result<reply::Response, warp::Rejection> {
-    Ok(apikit::reply::json(&VersionResponse {
-        version: VERSION.to_string(),
-    }))
+    Ok(apikit::reply::json(&VersionResponse::new(VERSION)))
 }
