+++ bin/menmosd/src/menmosd/server/handlers/storage/list.rs
@@ -1,6 +1,6 @@
 use apikit::{auth::UserIdentity, reject::InternalServerError};
 
-use interface::ListStorageNodesResponse;
+use protocol::directory::storage::ListStorageNodesResponse;
 
 use warp::reply;
