+++ crates/interface/src/lib.rs
@@ -1,5 +1,4 @@
 mod directory;
-pub mod message;
 mod storage;
 
 pub use directory::*;
