+++ crates/menmos-client/Cargo.toml
@@ -17,6 +17,7 @@ hyper = "0.14"
 mpart-async = "0.5"
 interface = {path = "../interface"}
 log = "0.4"
+protocol = {path = "../protocol"}
 reqwest = {version = "0.11", features = ["json"]}
 serde = {version = "1", features = ["derive"]}
 serde_json = "1"
