+++ crates/menmos-client/src/lib.rs
@@ -6,9 +6,9 @@ pub mod profile;
 use builder::ClientBuilder;
 use parameters::Parameters;
 
-pub use interface::message::directory_node::Query;
+pub use client::Client;
 pub use interface::BlobMeta as Meta;
+pub use interface::Query;
 pub use interface::QueryResponse;
 pub use interface::Type;
-pub use client::Client;
 pub use profile::Config;
