+++ crates/protocol/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "protocol"
+version = "0.1.0"
+authors = ["William Dussault <dalloriam@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+interface = {path = "../interface"}
+rapidquery = {path = "../rapidquery"}
+serde = {version = "1", features = ["derive"]}
